refactor(assert,expect): use shared `format.ts` and `diff.ts`
