+++ assert/assert_strict_equals.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
-import { format } from "./_format.ts";
+import { format } from "../internal/format.ts";
 import { AssertionError } from "./assertion_error.ts";
-import { buildMessage, diff, diffstr } from "./_diff.ts";
+import { buildMessage, diff, diffstr } from "../internal/diff.ts";
 import { CAN_NOT_DISPLAY } from "./_constants.ts";
 import { red } from "../fmt/colors.ts";
