+++ internal/diff.ts
