+++ internal/format.ts
@@ -1,8 +1,6 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
-// This file been copied to `std/expect`.
-
 /**
  * Converts the input into a string. Objects, Sets and Maps are sorted so as to
  * make tests less flaky
