+++ internal/format_test.ts
@@ -1,9 +1,9 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 import { green, red, stripAnsiCode } from "../fmt/colors.ts";
 import { assertEquals, assertThrows } from "../assert/mod.ts";
-import { format } from "./_format.ts";
+import { format } from "./format.ts";
 
-// This file is copied from `std/assert`.
+// This file been copied to `std/expect`.
 
 Deno.test("format() generates correct diffs for strings", () => {
   assertThrows(
