+++ internal/mod.ts
@@ -1,0 +1,9 @@
+// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
+
+/**
+ * Internal utilities for the public API of the Deno Standard Library.
+ *
+ * Note: this module is for internal use only and should not be used directly.
+ *
+ * @module
+ */
